requirements.txt
amulet-core
aiohttp
